+++ src/App.js
@@ -1,38 +1,73 @@
 import React from 'react'
 import {Route, Switch, withRouter, Link} from "react-router-dom"
-import {Trans, withTranslation} from 'react-i18next'
-import LanguageChooser from './LanguageChooser'
-import TestStoreContext from './TestStoreContext'
-import NotFound from './NotFound'
+
+import ProfileIcon from "@material-ui/icons/PersonOutlineOutlined"
+import DashboardIcon from "@material-ui/icons/DashboardOutlined"
+
+import {withTheme, BottomNavigation, BottomNavigationAction} from '@material-ui/core'
+
 import {routes} from './lib/router'
-import {Typography, Button} from '@material-ui/core'
-import Landing from './Landing'
 
-import OfflineStatus from './components/OfflineStatus'
+import {
+  Landing,
+  Profile,
+  Register,
+  OfflineStatus,
+  Dashboard,
+  NotFound,
+  Dialog
+} from './components'
+import {withTranslation} from 'react-i18next'
 
-const App = ({t}) =>
-  <div className="App">
+
+export const App = ({theme: {palette: {type}}}) =>
+  <div className="app" style={{backgroundColor: type === "dark" ? "#000" : ""}}>
     <Switch>
       <Route component={Landing} exact path={routes.ROOT}/>
-      <Route component={() =>
-        <header className="App-header">
-          <Button color="primary" component={Link} to="404" variant="outlined">
-            {t("not-found.go-to")}
-          </Button>
-          <TestStoreContext/>
-          <Typography variant="h3">
-            {t('welcome.title', {framework: "react-i18next"})}
-          </Typography>
-          <Trans i18nKey="welcome.intro">
-            To get started, edit <code>src/App.js</code> and save to reload.
-          </Trans>
-          <LanguageChooser/>
-        </header>
-      } exact path={routes.TEST}
-      />
+      <Route component={Register} exact path={routes.REGISTER}/>
+      <Route component={Profile} exact path={routes.PROFILE}/>
+      <Route component={Dashboard} exact path={routes.DASHBOARD}/>
       <Route component={NotFound}/>
     </Switch>
+    <Route
+      path="/"
+      render={({location: {pathname}}) => pathname !== "/" ? <Navigation/> : null}
+    />
     <OfflineStatus/>
+    <Dialog/>
   </div>
 
-export default withRouter(withTranslation('common')(App))
+export default withRouter(withTheme()(App))
+
+
+const navigation = [
+  {
+    id: "dashboard",
+    icon: <DashboardIcon/>,
+    to: routes.DASHBOARD
+  },
+  {
+    id: "profile",
+    icon: <ProfileIcon/>,
+    to: routes.PROFILE
+  }
+]
+
+export const Navigation = withTranslation("common")(withRouter(
+  ({t, location: {pathname}}) =>
+    <BottomNavigation
+      style={{position: "fixed", bottom: 0, width: "100vw"}}
+      value={pathname.replace("/", "")}
+    >
+      {navigation.map(({id, icon, to}) =>
+        <BottomNavigationAction
+          component={Link}
+          icon={icon}
+          key={id}
+          label={t(`navigation.${id}`)}
+          to={to}
+          value={id}
+        />
+      )}
+    </BottomNavigation>
+))