+++ src/components/Dashboard.jsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import {withTranslation} from 'react-i18next'
+import {withPage} from './shared'
+
+
+const Dashboard = () => <div></div>
+
+export default withTranslation("pages")(withPage(Dashboard, {namespace: "dashboard"}))