+++ src/components/Profile/LanguageChooser.jsx
@@ -1,0 +1,34 @@
+import React from 'react'
+import {Select, MenuItem, Typography} from "@material-ui/core"
+import {withTranslation} from "react-i18next"
+import {languages} from "../../locales/locales.json"
+
+const LanguageChooser = ({i18n}) => {
+
+  const changeLanguage = ({target: {value}}) => {
+    localStorage.setItem("preferredLanguage", value)
+    i18n.changeLanguage(value)
+  }
+
+  return(
+    <Select
+      color="secondary"
+      onChange={changeLanguage}
+      value={i18n.language}
+    >
+      {languages.map(({code, title, flag}) =>
+        <MenuItem
+          key={code}
+          selected
+          value={code}
+        >
+          <Typography variant="subtitle1">
+            <span aria-label={`{title} flag`} role="img" style={{paddingLeft: 4}}>{flag}</span> {title}
+          </Typography>
+        </MenuItem>
+      )}
+    </Select>
+  )
+}
+
+export default withTranslation("profile")(LanguageChooser)