+++ src/lib/router.js
@@ -1,5 +1,8 @@
 export const routes = {
   ROOT: "/",
+  DASHBOARD: "/dashboard",
   TEST: "/test",
-  FORGOT_PASSWORD: "/new-password"
+  FORGOT_PASSWORD: "/new-password",
+  REGISTER: "/register",
+  PROFILE: "/profile"
 }