+++ CHANGELOG.md
@@ -4,6 +4,31 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 
+## [1.0.0-beta1] - 2019-02-11
+### Added
+- bottom navigation
+- use `localStorage` to save preferred language over sessions
+- unified settings & profile page
+- dark mode
+    - use `localStorage` to save preferred dark/light mode over sessions
+- Dialog
+- DestructButton
+- authentication:
+    - registration
+    - login
+    - logout
+    - user deletion
+    - redirects:
+        - logged out
+            - after registering a user -> dashboard
+            - dashboard -> root
+            - 404 -> root
+### Changed
+- Offline status
+- UI element sizes increased
+### Removed
+- example codes
+
 ## [0.3.0] - 2019-01-30
 ### Added
 - logo